% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{anova1_wrap}
\alias{anova1_wrap}
\title{Conduct anova test}
\usage{
anova1_wrap(x, lhs, rhs, paracode)
}
\arguments{
\item{lhs}{A character string which is name of the main variable of numeric type}

\item{rhs}{A character string which is name of the group variable of factor type with multiple levels}

\item{paracode}{A character vector which is used for the specification of the parameter na.action.}

\item{x0}{dataframe}
}
\value{
an R List which contains 4 dataframes;
anova1_type1, anova1_type2, anova1_type3 and anova1_equalVariance_check.
}
\description{
This function immplements aov and Anova in car package after checking that each group in the input dataset has at least 2 non-null records
and at least 1 record for each level of the factor of the group variable.
}
\examples{
anova1_wrap(x, "BMI", "AgeCat", c("na.omit"))

}
