% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkUnique.R
\name{checkUnique}
\alias{checkUnique}
\title{Check the uniqueness for the subset for each realization of the variables}
\usage{
checkUnique(x, listStrings, type, text)
}
\arguments{
\item{x}{dataframe}

\item{listStrings}{A character vector which are names of variables}

\item{type}{A numeric string: 1 error (stop the transaction), 2 warning (warning a message)}

\item{text}{A character string which is a message printed out}
}
\description{
This function check that there is only a single record for the subset corresponding to each relization of the variables
}
\examples{
\dontrun{
checkUnique(demo, c("BMI", "SEX"), 1, "There exist multiple records for each realisation of the variables.")
}

}
