% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chisq_wrap.R
\name{chisq_wrap}
\alias{chisq_wrap}
\title{Conduct chi-square test}
\usage{
chisq_wrap(x0, lhs, rhs, correct0)
}
\arguments{
\item{x0}{dataframe}

\item{lhs}{A character string which is name of the main variable of factor type}

\item{rhs}{A character string which is name of the group variable of factor type}

\item{correct0}{A character vector which is used for the specification of the parameter correct.}
}
\value{
an R List which contains 3 dataframes;
chitest_result, chitest_count and chitest_check.
}
\description{
This function immplements chisq.test after checking that the number of the row and the number of the column of the contingency table
of the input dataset are both at least 2.
}
\examples{
\dontrun{
chisq_wrap(demo, "Sex", "AgeCat", "TRUE")
}

}
