% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_mp.R
\name{data_mp}
\alias{data_mp}
\title{For each realization of the variables (listStrings), call the stats_mp function for the variable (mainString)}
\usage{
data_mp(
  dsgroup,
  listStrings,
  mainString,
  type,
  statString,
  narmString,
  digitString,
  digitString2
)
}
\arguments{
\item{dsgroup}{dataframe}

\item{listStrings}{A character vector of the names of variables}

\item{mainString}{a numeric string: the name of the variable}

\item{type}{a numeric string: 1 Type 1 for Base dataset, 2 Type 2 for the output dataset of Stats or ML node}

\item{statString}{a numeric string: choose from (N, sum, mean, median, min, max, mode, sd, se, meanSD, meanSE)}

\item{narmString}{a character string: TRUE or FALSE for na.rm parameter}

\item{digitString}{a numeric string: the minimum number of digits to the right of the decimal point}

\item{digitString2}{a numeric string: the minimum number of digits to the right of the decimal point for SD or SE in meanSD and meanSE}
}
\value{
a character string
}
\description{
For each realization of the variables (listStrings),
apply the stats_mp function to the variable (mainString).
}
\examples{
\dontrun{
data_mp(demo, c("Treatment", "Sex", "Ethnic"), BMI, 1, "meanSE", TRUE, 1, 2)
}

}
