% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datacen.R
\name{datacen}
\alias{datacen}
\title{Normalize the numeric variable}
\usage{
datacen(dsgroup, lhs, rhs)
}
\arguments{
\item{dsgroup}{dataframe}

\item{lhs}{A character string which is name of the main variable of numeric type}

\item{rhs}{A character string which is name of the group variable of factor type}
}
\value{
the normalized variable
}
\description{
This function normalizes the numeric variable by computing:
the normalized variable = the numeric variable (lhs) - the mean of the numeric variable (lhs) for each value of the group variable (rhs)
}
\examples{
\dontrun{
datacen(demo, "BMI", "Sex")
}

}
