% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dstring.R
\name{dstring}
\alias{dstring}
\title{Obtain the actual variables}
\usage{
dstring(lstrings)
}
\arguments{
\item{lstrings}{A character vector of the names of variables
In the nested function x is a parameter for a dataframe.}
}
\value{
This function returns an R List whose objects are the columns in the dataframe;
}
\description{
From a list of the names of the variables in a dataset,
this function returns the actual variables in the dataset.
}
\examples{
\dontrun{
listV <- dstring(lstrings)
listVars <- listV(demo)
}

}
