% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_rbf.R
\name{f_rbf}
\alias{f_rbf}
\title{Run the regression_com for the evaluation of a subset of columns}
\usage{
f_rbf(
  rindex,
  reg_param,
  ite_n,
  trainlabel,
  traindata,
  testlabel,
  testdata,
  start,
  c0type
)
}
\arguments{
\item{rindex}{An integer vector of 0 and 1, which shows presence/absence of each column of the training dataset. (1 presence, 0 absence)}

\item{reg_param}{A double: the regularization parametrer}

\item{ite_n}{An integer: the iteration size in the inner loop}

\item{trainlabel}{A matrix: 1 x n1 matrix of training label where n1 is the number of records.}

\item{traindata}{A matrix: m x n1 matrix of training data where m is the number of variables and n1 is the number of records.}

\item{testlabel}{A matrix: 1 x n2 matrix of test label where n2 is the number of records.}

\item{testdata}{A matrix: m x n2 matrix of test data where m is the number of variables and n2 is the number of records.}

\item{start}{An integer vector which shows the original position of the variables. It is returned by the modify_spl function.}

\item{c0type}{An integer: 1 for classification, 2 for regression}
}
\description{
Run the regression_com function using a subset of the columns of the dataset
}
\examples{
\dontrun{
f_rbf(c(0, 1, 1, 0, 1), 1e-5, 3, trainlabel, traindata, testlabel, testdata, 10, c(1, 2, 3, 4, 6, 10), 1)
}

}
