% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{kruskal2_wrap}
\alias{kruskal2_wrap}
\title{Conduct Kruskal-Wallis test for the ordered factor variable}
\usage{
kruskal2_wrap(x, lhs, rhs, naaction)
}
\arguments{
\item{x}{dataframe}

\item{lhs}{A character string which is name of the main variable of ordered factor type}

\item{rhs}{A character string which is name of the group variable of factor type}

\item{naaction}{A character vector which is used for the specification of the parameter na.action.}
}
\value{
an R List which contains 2 dataframes;
kruskal_result and kruskal_count.
}
\description{
This function immplements kruskal.test after checking that each group in the input dataset has at least 2 non-null records
and at least 1 record for each level of the factor of the group variable.
}
\examples{
kruskal2_wrap(demo, "WeightCat", "AgeCat", "na.omit")

}
