% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kruskal_wrap.R
\name{kruskal_wrap}
\alias{kruskal_wrap}
\title{Conduct Kruskal-Wallis test for the numeric variable}
\usage{
kruskal_wrap(x, lhs, rhs, naaction)
}
\arguments{
\item{x}{dataframe}

\item{lhs}{A character string which is name of the main variable of numeric type}

\item{rhs}{A character string which is name of the group variable of factor type}

\item{naaction}{A character vector which is used for the specification of the parameter na.action.}
}
\value{
an R List which contains 1 dataframe;
kruskal_result.
}
\description{
This function immplements kruskal.test after checking that each group in the input dataset has at least 2 non-null records
and at least 1 record for each level of the factor of the group variable.
}
\examples{
\dontrun{
kruskal_wrap(demo, "BMI", "AgeCat", "na.omit")
}

}
