% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mat2data.R
\name{mat2data}
\alias{mat2data}
\title{convert a matrix which is the first object of a list to a dataframe}
\usage{
mat2data(idata)
}
\arguments{
\item{idata}{list}
}
\value{
dataframe
}
\description{
convert a matrix which is the first object of a list to a dataframe
}
\examples{
\dontrun{
mat2data(adj)
}

}
