% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miss_main.R
\name{miss_main}
\alias{miss_main}
\title{Create a dataframe with no missing values}
\usage{
miss_main(dsgroup, xNames, yName, vec, missType)
}
\arguments{
\item{dsgroup}{dataframe: input dataframe}

\item{xNames}{A string vector: the names of the input variables}

\item{yName}{A string: the name of the target variable}

\item{vec}{An integer vector (size of xNames): type of input variables (1 numerical variable, 2 ordered factor variable, 3 un-ordered factor variable)}

\item{missType}{An integer: 1 To remove the records with any missing values,
2 to replace the missing values with the mean (for the continuous variable) and the mode (for the categorical variable)}
}
\value{
dataframe of the input variables and the target variable with no missing values
}
\description{
Run the classification and the regresson methods in
"Scalable Kernel Methods via Doubly Stochastic Gradients".
}
\examples{
\dontrun{
miss_main(demo,c("BSSYS","BSDIA","BLWT","BLHT","AGECAT","WT2CLS","SEX","TRTGRP","ELIG","CRCLCT"), c("BMI"), c(1,1,1,1,2,2,3,3,3,3), 1)
}

}
