% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_main.R
\name{ml_main}
\alias{ml_main}
\title{Implement classification and regresson}
\usage{
ml_main(dsgroup, xNames, yName, vec, svmType, tSize, reg_param0, pSize)
}
\arguments{
\item{dsgroup}{dataframe: input dataframe}

\item{xNames}{A string vector: the names of the input variables}

\item{yName}{A string: the name of the target variable}

\item{vec}{An integer vector (size of xNames): type of input variables (1 numerical variable, 2 ordered factor variable, 3 un-ordered factor variable)}

\item{svmType}{An integer: 1 classification (SVM), 2 regression (SVM), 3 classification (MKL), 4 regression (MKL), 5 classification (PSO), 6 regression (PSO),}

\item{tSize}{An integer vector: (the percent for the training dataset, the percent for the test dataset, the percent for the validation dataset)}

\item{reg_param0}{A double vector: the regularization parametrer vector; if size > 1, the best parameter is searched using the validation dataset.}

\item{pSize}{An integer vector: (Population of PSO, Iteration number (PSO, MKL), Iteration number (SVM), p-norm of MKL, The best N Feature Subsets (PSO))}
}
\value{
dataframe of the input variables and the target variable with no missing values
}
\description{
Implement classification and regresson with SVM, MKL and PSO
}
\examples{
\dontrun{
ml_main(demo,c("Bssys","Bsdia","Blwt","Blht","AgeCat","Wt2Cls","Sex","treatment","Elig","Crclct"), c("BMI"), c(1,1,1,1,2,2,3,3,3,3), 2, 
c(70, 30, 0), c("1e-5"), c(NA, NA, 7, NA, NA))
}

}
