% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_spl.R
\name{modify_spl}
\alias{modify_spl}
\title{Delete redundant columns}
\usage{
modify_spl(train0, test0, var0, start0, start1, start2)
}
\arguments{
\item{train0}{dataframe: training dataset}

\item{test0}{dataframe: test dataset}

\item{var0}{dataframe: validation dataset}

\item{start0}{An integer vector: start0 returned by spl_factor}

\item{start1}{An string vector: start1 returned by spl_factor}

\item{start2}{An integer vector: start2 returned by spl_factor}
}
\value{
R List with the following 6 objects
\enumerate{
\item train1 dataframe: training dataset after deleting the columns
\item test1 dataframe: test dataset after deleting the columns
\item var1 dataframe: validation dataset after deleting the columns
\item olabel a string vactor: names of the original column names with no duplication
\item odimno a integer: the size of olabel
\item start an integer vactor: a vector which shows the starting position of the original columns
It is constructed as follows: starting with the integer vector c(1), and adding the size of each variable expanded by the one-hot Encoding Scheme.
}
}
\description{
Start with i = 1 and an integer vactor Del with size = 0;
if all rows of the column i of the training dataset are identical, append i to Del.
if (i > 1) and the rank of the training dataset from the column 1 to the column i - 1 and the rank of the training dataset from the column 1 to the column i are same,
append i to Del.
Delete the columns in Del of training dataset, test dataset and validation dataset.
}
\examples{
\dontrun{ 
modify_spl(traindata0,testdata0,vardata0,c(1,1,1,1,1,1,1,1,0,0,0,0,0,0,0),c("BSSYS","BSDIA","BLWT","BLHT","AGECAT","WT2CLS","SEX","TRTGRP","ELIG","ELIG","ELIG","ELIG","CRCLCT","CRCLCT","CRCLCT"),c("BSSYS","BSDIA","BLWT","BLHT","AGECAT","WT2CLS","SEX","TRTGRP","ELIG___1","ELIG___2","ELIG___3","ELIG___4","CRCLCT___1","CRCLCT___2","CRCLCT___3"))
}
 
}
