% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mregression_com.R
\name{mregression_com}
\alias{mregression_com}
\title{Run the MKL method (inner loop)}
\usage{
mregression_com(
  gk,
  gen_n,
  theta,
  p0,
  reg_param,
  ite_n,
  trainlabel,
  traindata,
  testlabel,
  testdata,
  odimno,
  start,
  c0type
)
}
\arguments{
\item{gk}{An integer: the current iteration number in the outer loop}

\item{gen_n}{An integer: the iteration size in the outer loop}

\item{theta}{A numeric vector: the weight of the kernel}

\item{p0}{A double: The p-norm which is a costraint on the coefficients of the kernels}

\item{reg_param}{A double: the regularization parametrer}

\item{ite_n}{An integer: the iteration size in the inner loop}

\item{trainlabel}{A matrix: 1 x n1 matrix of training label where n1 is the number of records.}

\item{traindata}{A matrix: m x n1 matrix of training data where m is the number of variables and n1 is the number of records.}

\item{testlabel}{A matrix: 1 x n2 matrix of test label where n2 is the number of records.}

\item{testdata}{A matrix: m x n2 matrix of test data where m is the number of variables and n2 is the number of records.}

\item{odimno}{An integer: the size of the original variables}

\item{start}{An integer vector which shows the original position of the variables. It is returned by the modify_spl function.}

\item{c0type}{An integer: 1 for classification, 2 for regression}
}
\description{
Run the inner loop of the MKL method in
"Simple and Efficient Multiple Kernel Learning by Group Lasso"
based on the classification and the regresson methods in
"Scalable Kernel Methods via Doubly Stochastic Gradients".
}
\examples{
\dontrun{
mregression_com(2, 3, c(0.1667, 0.1667, 0.1667, 0.1667, 0.1667, 0.1667), 2, 1e-5, 3, trainlabel, traindata, testlabel, testdata, 10, c(1, 2, 3, 4, 6, 10), 1)
}

}
