% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{mymode}
\alias{mymode}
\title{Comput mode}
\usage{
mymode(v)
}
\arguments{
\item{x}{vector}

\item{na.rm}{a character string (TRUE or FALSE) for the parameter na.rm}
}
\value{
mode
}
\description{
Comput mode which is the most frequent of all values of a input vector
}
\examples{
mymode(x, na.rm=TRUE)
}
