% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mystrsplit.R
\name{mystrsplit}
\alias{mystrsplit}
\title{wrap strsplit function}
\usage{
mystrsplit(string, dlm)
}
\arguments{
\item{string}{a character string:}

\item{dlm}{a character string: delimiter string}
}
\value{
a character vector
}
\description{
wrap strsplit function
In strsplit function if the string after the last delimiter string is "",
the character array does not include the last string.
This function returns the character array whose the last entry is "".
}
\examples{
\dontrun{
mystrsplit("There exist multiple records for each realisation of the variables.", " ")
}

}
