% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/out_figure.R
\name{out_figure}
\alias{out_figure}
\title{output the layered figures}
\usage{
out_figure(
  dsgroup,
  listName,
  stratStrings,
  checkStrings,
  type,
  path,
  levelpath,
  .fun = NULL,
  ...
)
}
\arguments{
\item{dsgroup}{dataframe}

\item{listName}{A character string for the name of the file which contains all realization of the layered variables}

\item{stratStrings}{A character vector eash of which is name of the layered variables}

\item{checkStrings}{A character vector eash of which is name of the check variables}

\item{type}{a numeric string: 1 Type 1 for Base dataset, 2 Type 2 for the output dataset of the Stats or the ML node}

\item{path}{A character string which is the path of the directory in which figures for each realization of the layered variables are output.}

\item{levelpath}{A character string which is the path of the directory in which levels of each layered variables are output.}

\item{.fun}{A function which is immplemented for the subset for the each realization of the layered variables}

\item{...}{other arguments passed on to .fun}
}
\description{
This function splits the dataset for each realization of the layered variables,
we immplement a function given by .fun for each subset and output a figure.
}
\examples{
\dontrun{
out_figure(demo, "demolist1", c("Ethnic", "Sex"), c("Agecat", "Treatment"), 1, 
"/home/shinichi/workspace/Repository/___project/test/Tree/testTree/BMI/TRTGRP/SEX/COUNTRY/barplot/___show",
"/home/shinichi/workspace/Repository/___project/test/Tree/testTree/BMI/TRTGRP/SEX/COUNTRY/barplot/___info/___levels", .fun = function(xx, yy) {
ggplot0 <- ggplot(data=xx, aes(x=Agecat, y=stat(count), fill=Treatment)) + 
geom_bar(stat="count", position=position_stack()) + 
labs(title ="Bar Plot of stat(count) based on Agecat", subtitle = yy) + 
theme(plot.title = element_text(hjust = 0.5)) + 
theme(plot.subtitle = element_text(hjust = 0.5))
return(ggplot0)
})
}

}
