% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{out_list}
\alias{out_list}
\title{output the layered lists}
\usage{
out_list(dsgroup, listName, stratStrings, listStrings, path, levelpath)
}
\arguments{
\item{dsgroup}{dataframe}

\item{listName}{A character string for the name of the file which contains all realization of the layered variables}

\item{stratStrings}{A character vector eash of which is name of the layered variables}

\item{listStrings}{A character vector eash of which is name of the input variables in the list}

\item{path}{A character string which is the path of the directory in which lists for each realization of the layered variables are output.}

\item{levelpath}{A character string which is the path of the directory in which levels of each layered variables are output.}
}
\description{
This function splits the dataset for each realization of the layered variables,
we output a list for each subset.
}
\examples{
out_list(demo, "demolist1", c("Treatment", "Sex"), c("Agecat", "Ethnic", "BMI", "Weight", "Height"), 2, 
"/home/shinichi/workspace/Repository/___project/test/Tree/testTree/BMI/TRTGRP/SEX/COUNTRY/anova/anova_type3/table011/___show",
"/home/shinichi/workspace/Repository/___project/test/Tree/testTree/BMI/TRTGRP/SEX/COUNTRY/anova/anova_type3/table011/___info/___levels")

}
