% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/out_table.R
\name{out_table}
\alias{out_table}
\title{output the layered tables}
\usage{
out_table(
  dsgroup,
  tableName,
  mainString,
  stratStrings,
  colStrings,
  rowStrings,
  type,
  statString,
  narmString,
  digitString,
  digitString2,
  path,
  levelpath
)
}
\arguments{
\item{dsgroup}{dataframe}

\item{tableName}{A character string for the name of the file which contains all realization of the layered variables}

\item{mainString}{A character string for the name of the variable for each cell in the tables}

\item{stratStrings}{A character vector eash of which is name of the layered variables}

\item{colStrings}{A character vector eash of which is name of the column variables}

\item{rowStrings}{A character vector eash of which is name of the row variables}

\item{type}{a numeric string: 1 Type 1 for Base dataset, 2 Type 2 for the output dataset of the Stats or the ML node}

\item{statString}{a numeric string: choose from (N, sum, mean, median, min, max, mode, sd, se, meanSD, meanSE)}

\item{narmString}{a character string: TRUE or FALSE for na.rm parameter}

\item{digitString}{a numeric string: the minimum number of digits to the right of the decimal point}

\item{digitString2}{a numeric string: the minimum number of digits to the right of the decimal point for SD or SE in meanSD and meanSE}

\item{path}{A character string which is the path of the directory in which tables for each realization of the layered variables are output.}

\item{levelpath}{A character string which is the path of the directory in which levels of each layered variables are output.}
}
\description{
This function splits the dataset for each realization of the layered variables and
output tables for the subsets.
}
\examples{
\dontrun{
out_table(demo, "demotable1", "BMI", c("Ethnic", "Sex"), c("AgeCat", "Treatment"), c("Elig"), 1, "meanSE", TRUE, 1, 2, 
"/home/shinichi/workspace/Repository/___project/test/Tree/testTree/BMI/Treatment/Sex/Country/table/___show",
"/home/shinichi/workspace/Repository/___project/test/Tree/testTree/BMI/Treatment/Sex/Country/table/___info/___levels")
}

}
