% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{pttest_wrap}
\alias{pttest_wrap}
\title{Conduct paired t test}
\usage{
pttest_wrap(x0, lhs, rhs, paracode)
}
\arguments{
\item{x0}{dataframe}

\item{lhs}{A character string which is name of the main variable of numeric type}

\item{rhs}{A character string which is name of the group variable of factor type with two levels}

\item{paracode}{A character vector which is used for the specification of the 4 parameters (alternative, var.equal, conf.level, na.action).}
}
\value{
an R List which contains one dataframe;
pttest_result.
}
\description{
This function immplements t.test with the parameter paired = TRUE after checking that each group in the input dataset has at least 2 non-null records.
}
\examples{
pttest_wrap(x, "BMI", "Sex", c("two.sided", "TRUE", "0.95", "na.omit"))

}
