% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save2data.R
\name{save2data}
\alias{save2data}
\title{Save Dataset}
\usage{
save2data(idata, sdata, path)
}
\arguments{
\item{idata}{dataframe}

\item{sdata}{A character string which is name of the dataframe}

\item{path}{A character string which is the path of the root directory under which the dataframe is saved}
}
\description{
call save2show, save2info, save2rds functions
}
\examples{
\dontrun{
save2data(adj, "adj", "/home/name/workspace/Repository/___project/test/Datasource/csv1/adj")
}

}
