% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save2rds.R
\name{save2rds}
\alias{save2rds}
\title{Save R Dataset}
\usage{
save2rds(idata, sdata, path)
}
\arguments{
\item{idata}{dataframe}

\item{sdata}{A character string which is name of the dataframe}

\item{path}{A character string which is the path in which the dataframe is saved}
}
\description{
call saveRDS function and save the dataframe
}
\examples{
\dontrun{
save2rds(adj, "adj", "/home/name/workspace/Repository/___project/test/Datasource/csv1/adj")
}

}
