% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save2show.R
\name{save2show}
\alias{save2show}
\title{Save Dataset tsv file}
\usage{
save2show(idata, sdata, path, header = TRUE)
}
\arguments{
\item{idata}{dataframe}

\item{sdata}{A character string which is name of the dataframe}

\item{path}{A character string which is the path in which the tsv file of the dataframe is saved}

\item{header}{A logical value for the parameter col_names: default = TRUE}
}
\description{
call write_tsv function and save the tsv file of the dataframe
}
\examples{
\dontrun{
save2show(adj, "adj", "/home/name/workspace/Repository/___project/test/Datasource/csv1/adj")
}

}
