% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{save_data}
\alias{save_data}
\title{Save Dataset}
\usage{
save_data(idata, sdata, path)
}
\arguments{
\item{idata}{dataframe}

\item{sdata}{A character string which is name of the dataframe}

\item{path}{A character string which is the path of the root directory under which the dataframe is saved}
}
\description{
call save_show, save_info, save_rds functions
}
\examples{
save_data(adj, "adj", "/home/name/workspace/Repository/___project/test/Datasource/csv1/adj")

}
