% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{save_info}
\alias{save_info}
\title{Save Dataset Info}
\usage{
save_info(idata, sdata, path)
}
\arguments{
\item{idata}{dataframe}

\item{sdata}{A character string which is name of the dataframe}

\item{path}{A character string which is the path in which the dataframe info is saved}
}
\description{
call contents function and save the dataframe information and the level files for the factor variables
}
\examples{
save_info(adj, "adj", "/home/name/workspace/Repository/___project/test/Datasource/csv1/adj")

}
