% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{save_rds}
\alias{save_rds}
\title{Save R Dataset}
\usage{
save_rds(idata, sdata, path)
}
\arguments{
\item{idata}{dataframe}

\item{sdata}{A character string which is name of the dataframe}

\item{path}{A character string which is the path in which the dataframe is saved}
}
\description{
call saveRDS function and save the dataframe
}
\examples{
save_rds(adj, "adj", "/home/name/workspace/Repository/___project/test/Datasource/csv1/adj")

}
