% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{save_show}
\alias{save_show}
\title{Save Dataset tsv file}
\usage{
save_show(idata, sdata, path, header = TRUE)
}
\arguments{
\item{idata}{dataframe}

\item{sdata}{A character string which is name of the dataframe}

\item{path}{A character string which is the path in which the tsv file of the dataframe is saved}
}
\description{
call write_tsv function and save the tsv file of the dataframe
}
\examples{
save_show(adj, "adj", "/home/name/workspace/Repository/___project/test/Datasource/csv1/adj")

}
