% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_mat.R
\name{simple_mat}
\alias{simple_mat}
\title{convert matrix to dataframe}
\usage{
simple_mat(indata, inmat)
}
\arguments{
\item{indata}{dataframe}

\item{inmat}{matrix}
}
\value{
dataframe
}
\description{
convert matrix x to dataframe
If the name of each column of x is contained in the dataframe (indata),
it is converted to the same type of the variable in indata.
}
\examples{
\dontrun{
simple_mat(adj, inmat)
}

}
