% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spl_factor.R
\name{spl_factor}
\alias{spl_factor}
\title{One-hot Encoding Scheme}
\usage{
spl_factor(dsg0, vec)
}
\arguments{
\item{dsg0}{dataframe: input dataframe}

\item{vec}{An integer vector (size of xNames): type of input variables (1 numerical variable, 2 ordered factor variable, 3 un-ordered factor variable)}
}
\value{
R List with the following 4 objects
\enumerate{
\item dsg1 dataframe after One-hot Encoding Scheme
\item start0 a vactor (same size as the number of the columns of dsg1):
1 columns with no application of the one-hot Encoding Scheme, 0 columns created by the one-hot Encoding Scheme
\item start1 a vactor (same size as the number of the columns of dsg1): a vector with the names of the original columns
\item start2 a vactor (same size as the number of the columns of dsg1): a vector with the names of the new columns
}
}
\description{
Apply the one-hot Encoding Scheme to the un-ordered factor variables
}
\examples{
\dontrun{
spl_factor(demo, c(1,1,1,1,2,2,3,3,3,3))
}

}
