% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_mp.R
\name{stats_mp}
\alias{stats_mp}
\title{apply a summary statistic to a numeric vector and apply a format function to the result}
\usage{
stats_mp(vpl, type, statString, narmString, digitString, digitString2)
}
\arguments{
\item{vpl}{a numeric vector}

\item{type}{a numeric string: 1 Type 1 for Base dataset, 2 Type 2 for the output dataset of Stats or ML node}

\item{statString}{a numeric string: choose from (N, sum, mean, median, min, max, mode, sd, se, meanSD, meanSE)}

\item{narmString}{a character string: TRUE or FALSE for na.rm parameter}

\item{digitString}{a numeric string: the minimum number of digits to the right of the decimal point}

\item{digitString2}{a numeric string: the minimum number of digits to the right of the decimal point for SD or SE in meanSD and meanSE}
}
\value{
a character string: result
}
\description{
apply a summary statistic (N, sum, mean, median, min, max, mode, sd, se, meanSD, meanSE) to a numeric vector
and apply a round function and a format function to the result if type = 1.
check uniquness; then 1. apply a round and format function or 2. return the input value as it is if type = 2.
}
\examples{
\dontrun{
stats_mp(vector0, 1, "meanSE", TRUE, 1, 2)
}

}
