% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary0.R
\name{summary0}
\alias{summary0}
\title{Comput summary statistics}
\usage{
summary0(x, var, na.rm0)
}
\arguments{
\item{x}{dataframe}

\item{var}{a character string which is the name of a variable}

\item{na.rm0}{a character string (TRUE or FALSE) for the parameter na.rm}
}
\value{
the dataframeof summary statistics.
}
\description{
Create a dataframe whose variables are mean, sd, median, min, max, q1 and q3.
}
\examples{
\dontrun{
summary0(x, "Age", TRUE)
}

}
