% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_wrap.R
\name{summary_wrap}
\alias{summary_wrap}
\title{Compute summary statistics of a continuous variable}
\usage{
summary_wrap(x, var, na.rm0)
}
\arguments{
\item{x}{dataframe}

\item{var}{A character string which is name of the main variable of numeric type}

\item{na.rm0}{A character string which is used for the specification of the parameter na.rm.}
}
\value{
an R List which contains 1 dataframe;
summary_result.
}
\description{
Compute summary statistics of a continuous variable
}
\examples{
\dontrun{
summary_wrap(demo, "BMI", TRUE)
}

}
