% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryN_wrap.R
\name{summaryN_wrap}
\alias{summaryN_wrap}
\title{Compute summary statistics of a categorical variable}
\usage{
summaryN_wrap(x, var, exclude0)
}
\arguments{
\item{x}{dataframe}

\item{var}{A character string which is name of the main variable of numeric type}

\item{exclude0}{A character string which is used for the specification of the parameter exclude of the table function.}
}
\value{
an R List which contains 1 dataframe;
summary_result.
}
\description{
Compute summary statistics of a categorical variable
}
\examples{
\dontrun{
summaryN_wrap(demo, "AgeCat", "c(NA, NaN)")
}

}
