% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{survival_data}
\alias{survival_data}
\title{Create dataset for survival curve}
\usage{
survival_data(x0, lhs, rhs, rhs2, paracode)
}
\arguments{
\item{x0}{dataframe}

\item{lhs}{A character string which is name of the variable of numeric type for Time}

\item{rhs}{A character string which is name of the variable of numeric type for Event (1=target event, 0=censor)}

\item{rhs2}{A character string which is name of the group variable of factor type}

\item{naaction}{A character vector which is used for the specification of the parameter na.action.}
}
\value{
an R List which contains 2 dataframes;
survdat and survcen.
}
\description{
Create dataset for survival curve
}
\examples{
survival_data(x, "Time", "Event", "AgeCat", c("na.omit"))

}
