% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_wrap.R
\name{survival_wrap}
\alias{survival_wrap}
\title{Conduct survival analysis}
\usage{
survival_wrap(x0, lhs, rhs, rhs2, paracode)
}
\arguments{
\item{x0}{dataframe}

\item{lhs}{A character string which is name of the variable of numeric type for Time}

\item{rhs}{A character string which is name of the variable of numeric type for Event (1=target event, 0=censor)}

\item{rhs2}{A character string which is name of the group variable of factor type}

\item{paracode}{A character vector which is used for the specification of the parameter na.action.}
}
\value{
an R List which contains 2 dataframes;
survfit_result and survdiff_result.
}
\description{
Conduct survival analysis
}
\examples{
\dontrun{
survival_wrap(x, "Time", "Event", "AgeCat", c("na.omit"))
}

}
