% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_header.R
\name{table_header}
\alias{table_header}
\title{output the headers for the table}
\usage{
table_header(dp1, fName, colStrings, rowStrings, path)
}
\arguments{
\item{dp1}{dataframe}

\item{fName}{A character string for the name of the file}

\item{colStrings}{A character vector eash of which is name of the column variables}

\item{rowStrings}{A character vector eash of which is name of the row variables}

\item{path}{A character string which is the path of the directory in which the headers for the table are output.}
}
\description{
This function output the headers for the table
}
\examples{
\dontrun{
table_header(dp0, "All", c("Agecat", "Treatment"), c("Elig"),
"/home/shinichi/workspace/Repository/___project/test/Tree/testTree/BMI/TRTGRP/SEX/COUNTRY/table/___show")
}

}
