% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_strat_s1.R
\name{table_strat_s1}
\alias{table_strat_s1}
\title{output a dataframe which contains all realization of the variables}
\usage{
table_strat_s1(dsgroup, colStrings)
}
\arguments{
\item{dsgroup}{dataframe}

\item{colStrings}{A character vector eash of which is name of the variables}
}
\value{
a dataframe which contains all realization of the variables
}
\description{
This function generate a dataframe which contains all realization of the variables given by colStrings
}
\examples{
\dontrun{
table_strat_s1(demo, c("Ethnic", "Sex", "AgeCat", "Treatment"))
}

}
