% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ttest_wrap.R
\name{ttest_wrap}
\alias{ttest_wrap}
\title{Conduct t test}
\usage{
ttest_wrap(x0, lhs, rhs, paracode)
}
\arguments{
\item{x0}{dataframe}

\item{lhs}{A character string which is name of the main variable of numeric type}

\item{rhs}{A character string which is name of the group variable of factor type with two levels}

\item{paracode}{A character vector which is used for the specification of the 4 parameters (alternative, var.equal, conf.level, na.action).}
}
\value{
an R List which contains two dataframes;
ttest_result and ttest_equalVariance_check.
}
\description{
This function immplements t.test after checking that each group in the input dataset has at least 2 non-null records.
}
\examples{
\dontrun{
ttest_wrap(x, "BMI", "Sex", c("two.sided", "TRUE", "0.95", "na.omit"))
}

}
