% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateRDS.R
\name{updateRDS}
\alias{updateRDS}
\title{Update a dataframe}
\usage{
updateRDS(dsgroup, nameStrings)
}
\arguments{
\item{dsgroup}{dataframe}

\item{nameStrings}{A character vector of the names of the variables}
}
\value{
dataframe in which the variables not in a list of the names of variables are deleted
}
\description{
Update a dataframe by deleting the variables which are not in a list of the names of variables
}
\examples{
\dontrun{
updateRDS(adj, c("subjectID", "adjDate", "adjCode"))
}

}
