% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wil2_wrap.R
\name{wil2_wrap}
\alias{wil2_wrap}
\title{Conduct Wilcoxon Rank Sum test for ordered factor variable}
\usage{
wil2_wrap(x0, lhs, rhs, paracode)
}
\arguments{
\item{x0}{dataframe}

\item{lhs}{A character string which is name of the main variable of ordered factor type}

\item{rhs}{A character string which is name of the group variable of factor type with two levels}

\item{paracode}{A character vector which is used for the specification of the 4 parameters (alternative, correct, conf.level, na.action).}
}
\value{
an R List which contains two dataframes;
wil_ranksum_result and wil_ranksum_count.
}
\description{
This function immplements wilcox.test after checking that each group in the input dataset has at least 2 non-null records
}
\examples{
\dontrun{
wil2_wrap(demo, "BMI", "Sex", c("two.sided", "TRUE", "0.95", "na.omit"))
}

}
