% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{dstring}
\alias{dstring}
\title{Obtain the actual variables}
\usage{
dstring(lstrings)
}
\arguments{
\item{lstrings}{A character vector of the names of variables}

\item{x}{dataframe}
}
\value{
This function returns a list whose objects are the variables in the dataset;
}
\description{
From a list of the names of the variables in a dataset,
this function returns the actual variables in the dataset.
}
\examples{
listV <- dstring(lstrings)
listVars <- listV(x)

}
