% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/levelfile.R
\name{levelfile}
\alias{levelfile}
\title{Output the levels of the factor variables}
\usage{
levelfile(stratStrings, stratVars, levelpath)
}
\arguments{
\item{stratStrings}{A character vector of the names of variables}

\item{stratVars}{A list whose objects are the variables}

\item{levelpath}{A character string which is the path in which the levels of the factor variable is saved}
}
\description{
Output the levels of the factor variables in the directory of levelpath.
}
\examples{
\dontrun{
levelfile(lstrings, listVars, "/home/shinichi/workspace/Repository/___project/test/Tree/testTree/BMI/TRTGRP/SEX/anova/anova_type3/table11/___info/___levels")
}

}
