% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mymode.R
\name{mymode}
\alias{mymode}
\title{Comput mode}
\usage{
mymode(v)
}
\arguments{
\item{v}{vector}
}
\value{
mode
}
\description{
Comput mode which is the most frequent of all values of a input vector
}
\examples{
\dontrun{
mymode(v)
}

}
